Abstract ReportSidebar into SideNav component
